% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_int.R
\name{add_int}
\alias{add_int}
\alias{do_add_int}
\title{Add integer vectors}
\usage{
do_add_int(X, Y)
}
\arguments{
\item{X, Y}{Integer vectors representing the digits of big integers.
The first digit is the 1s column, second digit is tens column etc.}
}
\value{
Integer vector representing addition of \code{X} and \code{Y}.
}
\description{
Add integer vectors
}
\examples{
add_int(1, 1)
add_int(55, 56)
add_int(55, add_int(1, 55))

}
